[#30] feat: 그룹 가입 멤버 조회 기능 구현
- 그룹에 가입 되어있는 멤버들을 조회합니다.
- 그룹 멤버를 조회할 때 가입 상태를 필터링하여 멤버를 조회합니다.

+++ src/main/java/net/skhu/wassup/app/group/service/GroupInviteServiceImpl.java
@@ -1,6 +1,8 @@
 package net.skhu.wassup.app.group.service;
 
+import static net.skhu.wassup.global.error.ErrorCode.NOT_FOUND_GROUP;
 import static net.skhu.wassup.global.error.ErrorCode.NOT_FOUND_MEMBER;
+import static net.skhu.wassup.global.error.ErrorCode.UNAUTHORIZED_ADMIN;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -9,18 +11,13 @@ import net.skhu.wassup.app.group.domain.Group;
 import net.skhu.wassup.app.group.domain.GroupRepository;
 import net.skhu.wassup.app.member.domain.Member;
 import net.skhu.wassup.app.member.domain.MemberRepository;
-import net.skhu.wassup.global.error.ErrorCode;
 import net.skhu.wassup.global.error.exception.CustomException;
 import net.skhu.wassup.global.message.SMSMessageSender;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
 public class GroupInviteServiceImpl implements GroupInviteService {
-
-    private static final Logger log = LoggerFactory.getLogger(GroupInviteServiceImpl.class);
 
     private final SMSMessageSender smsMessageSender;
 
@@ -30,16 +27,16 @@ public class GroupInviteServiceImpl implements GroupInviteService {
 
     @Override
     @Transactional
-    public void send(Long groupId, RequestInviteGroup requestInviteGroup) {
+    public void send(Long adminId, Long groupId, RequestInviteGroup requestInviteGroup) {
         Group group = groupRepository.findById(groupId)
-                .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_GROUP));
+                .orElseThrow(() -> new CustomException(NOT_FOUND_GROUP));
+
+        validateAdmin(adminId, group);
 
         String code = group.getUniqueCode();
         String message = group.getName() + "/n" + requestInviteGroup.link() + "\n 초대 코드 : " + code;
 
         smsMessageSender.send(requestInviteGroup.phoneNumber(), "Wassup 초대 코드를 아래 링크에 입력해주세요.", message);
-
-        log.info("그룹 초대 코드 전송: Group={}, Phone={}, Code={}", group.getName(), requestInviteGroup.phoneNumber(), code);
     }
 
     @Override
@@ -48,12 +45,10 @@ public class GroupInviteServiceImpl implements GroupInviteService {
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new CustomException(NOT_FOUND_MEMBER));
 
-        Long groupId = member.getGroup().getId();
-        currentUserIsAdmin(id, groupId);
+        Group group = member.getGroup();
+        validateAdmin(id, group);
 
         member.accept();
-
-        log.info("그룹 초대 수락: Member={}", memberId);
     }
 
     @Override
@@ -62,20 +57,15 @@ public class GroupInviteServiceImpl implements GroupInviteService {
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new CustomException(NOT_FOUND_MEMBER));
 
-        Long groupId = member.getGroup().getId();
-        currentUserIsAdmin(id, groupId);
+        Group group = member.getGroup();
+        validateAdmin(id, group);
 
         memberRepository.deleteById(memberId);
-
-        log.info("그룹 초대 거절: Member={}", memberId);
     }
 
-    private void currentUserIsAdmin(Long id, Long groupId) {
-        Group group = groupRepository.findById(groupId)
-                .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_GROUP));
-
+    private void validateAdmin(Long id, Group group) {
         if (!group.getAdmin().getId().equals(id)) {
-            throw new CustomException(ErrorCode.UNAUTHORIZED_ADMIN);
+            throw new CustomException(UNAUTHORIZED_ADMIN);
         }
     }
